refactor(output): move to file
utils 패키지에서 viewer로 변경
